# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Haven LLM Studio API
  version: 0.1.0
  description: |
    OpenAI-compatible local LLM inference API.
    Supports completions, chat, streaming, embeddings, and model management.
servers:
  - url: http://127.0.0.1:1234
    description: Local default
  - url: http://localhost:1234
    description: Local alias
paths:
...
  /v1/completions:
    post:
      summary: Text completion
      operationId: createCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                model:
                  type: string
                stream:
                  type: boolean
                max_tokens:
                  type: integer
                temperature:
                  type: number
                top_p:
                  type: number
                top_k:
                  type: integer
                repeat_penalty:
                  type: number
                stop:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Completion result (or SSE stream if stream=true)
  /v1/chat/completions:
    post:
      summary: Chat completion
      operationId: createChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                        enum: [system, user, assistant]
                      content:
                        type: string
                model:
                  type: string
                stream:
                  type: boolean
                max_tokens:
                  type: integer
                temperature:
                  type: number
      responses:
        '200':
          description: Chat completion (or SSE stream if stream=true)
...